requirements.txt
pandas>=2,<3
tabulate>=0.9,<1.0
langchain>=1.0,<2.0
langchain_ollama>=1.0,<2.0
fastapi[standard]>=0.120,<0.130
mlflow>=3.5,<3.6
load_dotenv
html-to-markdown>=2.0,<3.0
hydra-core>=1.3.2,<1.4.0
mem0ai>=1.0,<1.1
protobuf>=5.29.5,<5.30.0
kuzu>=0.11,<0.12
rank-bm25>=0.2.2,<0.3.0